Create sharded PTT implementation
Summary:
This diff adds the function that creates the sharded type checking sub-sub-targets. They are accessible and runnable by building `<target_path>:<target>[typecheck][shard_<file_path>]`.

Upcoming diffs will update the BXL scripts that kick off the type check operations to allow sharded type check runs.
 ---

I'm going to be making type checking operations shardable, which should speed up type checking (especially for IDE use cases). The method of doing this will be to create sub-targets under the `typecheck` sub-target, all of which we will send to RE to calculate type errors.

The idea here is to kick off separate RE jobs for each file, where we type check only that file. All other files in the target (and all dependencies) are treated as dependencies, and are not type checked.

Followup: type check only the shards relating to the specific files that are open (for IDE use).

effective_sharding

Reviewed By: grievejia

Differential Revision: D64063901

fbshipit-source-id: 564b465b425bfeff1f4bc3ce06f2fb9b024f85ba