Refactor batched type check functionality
Summary:
This diff pulls out the batched type check runner into it's own function. The `check_targets` function now handles kicking off the run and outputting the result.

Upcoming diffs will add sharding functionality here.
 ---

I'm going to be making type checking operations shardable, which should speed up type checking (especially for IDE use cases). The method of doing this will be to create sub-targets under the `typecheck` sub-target, all of which we will send to RE to calculate type errors.

The idea here is to kick off separate RE jobs for each file, where we type check only that file. All other files in the target (and all dependencies) are treated as dependencies, and are not type checked.

Followup: type check only the shards relating to the specific files that are open (for IDE use).

effective_sharding

Reviewed By: grievejia

Differential Revision: D64063894

fbshipit-source-id: 21cfd8b5e4ba5a7dd30941c11651a8a3cc0f40de