Preserve ArtifactInfo when using selective_debugging.
Summary:
When focused debugging is enabled we "scrub" the binary and filter the `ArtifactTSet` to only include
a subset of target's debug artifacts. When we are done filtering we don't rebuild our tset, we make a single node that has all the children of all
the targets in the apple_binary.

This causes downstream failures to `debug_artifacts_validators` because the TSet not only looks different, but, the corresponding `.label`
to the artifacts has changed, which requires users to maintain multiple "allowlists" based on the build config.

Instead, preserve the entire `ArtifactInfo` for debug info we want to keep. This makes the `label` corresponding to debug artifacts stay the same
whether we are in focused debugging or not.

Reviewed By: chatura-atapattu

Differential Revision: D65238543

fbshipit-source-id: 60d9a6bb450e200369eab5bbcc017e3e90bd5b2d