Make vendored json into proper target
Summary:
This changes the vendored json module to be a proper target.
This is necessary, since in some places where toolchain applications using json were included in release or otherwise packaged, the json module would be missing. This solves the issue and simplifies internals removing the need to prepend the utility modules path all over the place.

This is first part - that we need to sync back before progressing and using the new application for testing

Reviewed By: TheGeorge

Differential Revision: D65274428

fbshipit-source-id: a92a1dc38cddbaa708093e9a5f8c72c5294b71cb