Fix dthin-lto premerger file magic failure
Summary:
During thin-link distributed build write out `.merged.bc` suffixed files for each input bitcode file. These files are
either LLVM bitcode files, empty files or ASCII text files. Each case communicates to Buck the results of the pre-merger
process for this particular input bitcode file.

In order to process each such file the dist_lto_planner_darwin wrapper Python script runs `file` on each file to determine if it
is LLVM bitcode and proceeds accordingly based on the result. We were experiencing crashes when `file` occassionally,
reported LLVM bitcode files to be of type "Berkeley DB" instead of the expected "LLVM bitcode, wrapper".

Here is an example failure: https://www.internalfb.com/sandcastle/workflow/1179943102375964366/artifact/actionlog.1179943102464878730.stdout.1?selectedLines=715-751-1-1

This seems to be because file magic numbers don't actually uniquely identify file types. File works by interpretting
a list of ordered conditions recorded in `/usr/share/misc/magic`. Each condition is associated with text to print identifying
the file type, and it takes the first condition matched.

First in the file, we have the Berkeley DB conditions:

```
# Berkeley DB
#
# Ian Darwin's file /etc/magic files: big/little-endian version.
#
# Hash 1.85/1.86 databases store metadata in network byte order.
# Btree 1.85/1.86 databases store the metadata in host byte order.
# Hash and Btree 2.X and later databases store the metadata in host byte order.

# ...

# From Max Bowsher.
12	long	0x00040988	Berkeley DB          # The condition we do match
>16	long	>0		(Log, version %d, native byte-order)
12	belong	0x00040988	Berkeley DB
>16	belong	>0		(Log, version %d, big-endian)
12	lelong	0x00040988	Berkeley DB
>16	lelong	>0		(Log, version %d, little-endian)
```

Then later we have the LLVM bitcode conditions:

```
#------------------------------------------------------------------------------
# $File: llvm,v 1.9 2019/04/19 00:42:27 christos Exp $
# llvm:  file(1) magic for LLVM byte-codes
# URL:  https://llvm.org/docs/BitCodeFormat.html
# From: Al Stone <ahs3@fc.hp.com>

0	string	llvm	LLVM byte-codes, uncompressed
0	string	llvc0	LLVM byte-codes, null compression
0	string	llvc1	LLVM byte-codes, gzip compression
0	string	llvc2	LLVM byte-codes, bzip2 compression

0	lelong	0x0b17c0de	LLVM bitcode, wrapper  # The condition we want to match
# Are these Mach-O ABI values?  They appear to be.
>16	lelong	0x01000007	x86_64
>16	lelong	0x00000007	i386
>16	lelong	0x00000012	ppc
>16	lelong	0x01000012	ppc64
>16	lelong 	0x0000000c	arm

0	string	BC\xc0\xde	LLVM IR bitcode
```

The condition format is a byte offset, then a type, then a literal to match against which is interpretted according to the preceeding type.
The first Berkeley DB conditions reads, check if 12 bytes into the file you find the following long interpreted with system endianness. The following LLVM bitcode condition reads, check if 0 bytes into the file you find the following long, but interpret the literal as little-endian.

Obviously since the offsets are different, both these conditions can be true, and the Berkeley DB condition takes precendence since it appears first in the file. That is exactly what happened with the problematic LLVM bitcode file. We see the LLVM bitcode magic in the first four bytes, but starting at 12 bytes in we see the Berkeley DB magic:

```
00000000: dec0 170b 0000 0000 1400 0000 8809 0400  ................
          ^LLVM magic                   ^ Berkeley DB magic
```

The docs https://llvm.org/docs/BitCodeFormat.html#bitcode-wrapper-format indicate that the bytes at that position indicate the size of the stream. If you just so happen to have a LLVM bitcode file of the right size, `file` will report it as a `Berkeley DB` file.

To fix this, I wrote my own predicate to just check the first four bytes since we only care if the file is LLVM bitcode or ASCII, there are no
other options.

Reviewed By: rmaz

Differential Revision: D65636990

fbshipit-source-id: 28bb05756d98b36ae88b336a3ed42b3ab68ca465