Add `toolchains//:test` for test toolchain
Summary:
One of usages of `read_config("fbcode", "sanitizer")` is populating the sanitizer label for test infra. Currently, the most viable option to remove this `read_config` is to populate this label at analysis time instead of at load time.

To do this, I'm starting by adding a `toolchains//:test` toolchain. This toolchain will eventually be a dep of every test rule. This toolchain currently just has the sanitizer attribute set to a select so to enable adding sanitizer label to `ExternalRunnerTestInfo` provider at analysis time. Eventually, we will expand to other labels that need to be cleaned up from using `read_config`.

Reviewed By: blackm00n

Differential Revision: D65574343

fbshipit-source-id: 7aa2c4865c3dc280f7849960d856aad435faf422