Back out "Back out "[buck2][python][prelude] Do not include live pars in standalone resources""
Summary:
Original commit changeset: 0bd14da72162
"
With inplace by default enabled if a python binary is added as a resource the link tree would included as a hidden_resource. This caused standalone builds to fail - since we do not allow hidden resources in standalone pars.
This diff checks the resources to see if we are including a python binary. If so we add the standalone output as resource.
We track default_resources so that when building an inplace par we can include resources as inplace.
https://fb.workplace.com/groups/python.builds/permalink/3888593271462086/
S461975
#buildmore
"

Reviewed By: cxxxs

Differential Revision: D65785835

fbshipit-source-id: c8c6cb6f7e7c2755f36e4e43cc9d83f7f58edef8