Enable used_classes for incremental builds
Summary:
[Doc](https://docs.google.com/document/d/1-zhGPkmt3I2os-kcIaFk9SvT6odt2AJWswSY-J_6AWs/edit?usp=sharing)

### Context
We are exploring the integration of the Kotlin Incremental Compiler into Buck for the incremental compilation of Kotlin source files. The goal is to improve build speed by only recompiling modified source files within a target, rather than rebuilding the entire target as is currently done.

We had to disable the used_classes plugin since it was incompatible with the incremental compilation mechanism. For more information, see D65209518.

### This diff
In a previous diff (D66009177), a basic support for an incremental compiler merging the previous kotlin_used_classes.json file with the new one was added.
This diff enabled used_classes for targets using incremental compiler.

Differential Revision: D66164293

fbshipit-source-id: 31401f511d7d55349dff69d5221b02357662edff