Turn on remote cache uploads for JavaCD actions
Summary:
These were turned on previously, but turned off in D63386662. That was because we saw a regression for WhatsApp Android CI builds, but that regression was mostly because we were doing uploads for anything with `allow_cache_upload = True` too, which was fixed in D63532344.

Let's have another go. Note that this will still not cause any remote dep-file cache hits to occur, since the code for checking for remote dep-files is only turned on for members of Build Infra.

Reviewed By: ezgicicek

Differential Revision: D66210577

fbshipit-source-id: 299bd92524d10813616c9cf18f7766f72937d3b1