Add buckconfig-backed modifiers
Summary:
To migrate off of modefiles, we need to migrate some amount of `--config` to modifiers and some amount of `read_config` to selects. Migrating `read_config` to selects is doable, but migrating `--config` to some set of CLI or in-repo modifiers thoroughly is basically impractical. This means that there's no safe way to switch over from a `read_config` to a `select` and ensure that behavior stays the same.

The alternative is currently to keep the `read_config` or use a select on buckconfig in the default arm of the select, but both of these solutions have issues in that they make it significantly harder for us to query for usages of reading buckconfig that are migrated and usages that haven't. This is especially problematic given that `read_config` requires dynamic probing to be cleaned thoroughly and thus we cannot know statically which buckconfigs we can get rid of. Additionally, use of additional nested select can increase starlark memory usage.

This diff introduces a different solution. We will add an explicit buckconfig to modifier mapping, such that if buckconfig is satisfied, the mapped modifiers are automatically added to the configuration. This enables us to always have equivalent modifiers specified for when `--config` is specified.

Reviewed By: samkevich

Differential Revision: D66048059

fbshipit-source-id: ad27f6dadcc0027eeef75fb81ddc1df8795d01c3