26/n: elp.bxl: Tag each target origin as either app, dep or prelude
Summary:
We now have the erlang prelude as a default target added in `elp.bxl`.  We also pass in a list of dependency targets via `deps_targets`.

We postprocess the output of `elp.bxl` in ELP, and want to make sure we do not load targets from the prelude and dependencies that we are not interested in for our specific target.

To enable this, we tag the origin of each target we return, as either being `app`, `dep`, or `prelude`.

Reviewed By: perehonchuk

Differential Revision: D66306312

fbshipit-source-id: d177ebc62a8def489fef999820ba33ad2092750f