Fix dist_lto_planner script
Summary:
In Python loops don't have their own scope and variables introduced by the loop can leak beyond the end of the loop. Here we
have a loop variable (path) that shadows (and ultimately overrides) and outer variable. The two are meant to be distinct path
variables, and the inner overidding the outer has horrible hard to diagnose consequences. Fix.

Reviewed By: jtbraun

Differential Revision: D66563629

fbshipit-source-id: 61e152e7bf22a0afa3903699eccf71a4ca802e74