Prelude: `apple_tools()` is a dep, not exec_dep
Summary:
**Problem**

`_apple_tools` is defined as an `exec_dep` and this is problematic - it means that any `select()` on the `apple_tools()` target's attributes will be resolved against the exec platform, not the target platform.

This is not consistent with how toolchains (e.g., `apple_toolchain()`, `cxx_toolchain()`) behave which are configured against the target platform. Specifically, it precludes the ability to pick different tools based on the target platform (which we want to take advantage of).

**Solution**

Make `_apple_tools` be a `dep` and ensure each tool *itself* is an `exec_dep`.

Reviewed By: IanChilds

Differential Revision: D66601523

fbshipit-source-id: 2d775747c685e4b85e7534c6b451b063977dfda0