Fix potential issue parsing epmd output
Summary:
# Context
When running an Erlang test, before starting the test-node (that will actually run the test), we start epmd on a port, and make sure it is up.

For the latter, we parse stdout/stderr and try to find a concrete message, that means that epmd is ready

# Problem
By default, a port will send packets of lines that not necessarily will contain full lines, so in theory, it is possible in some cases to miss the message (in which case, we should timout after 1 sec and retry

# This diff
To avoid issues, use the `line` option of `spawn_port()` to request that we get full lines. We don't expect epmd to send super long lines, so if this were to happen, we can error right away.

Reviewed By: TheGeorge, acw224

Differential Revision: D66759103

fbshipit-source-id: d90840c16ca5591928a47c018af897b1554e0131